[sandbox] rename js files to jsx
