+++ sandbox/src/OrgTwo.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import useSendPageViewEvent from "./useSendPageViewEvent";
+import useSendPageViewEvent from "./useSendPageViewEvent.js";
 
 export default function OrgTwo() {
   useSendPageViewEvent({ instanceName: "organizationTwo" });
