+++ sandbox/src/PersonalizationFormBased.jsx
@@ -1,6 +1,6 @@
 import React, { useEffect, useState } from "react";
-import ContentSecurityPolicy from "./components/ContentSecurityPolicy";
-import useSendPageViewEvent from "./useSendPageViewEvent";
+import ContentSecurityPolicy from "./components/ContentSecurityPolicy.jsx";
+import useSendPageViewEvent from "./useSendPageViewEvent.js";
 
 const SCOPES_FOR_PAGE = ["sandbox-personalization-page2"];
