+++ sandbox/src/components/ContentSecurityPolicy.jsx
