+++ sandbox/src/components/Cookies.jsx
