+++ sandbox/src/components/UnsafeContentSecurityPolicy.jsx
