+++ sandbox/src/index.jsx
@@ -12,8 +12,8 @@ governing permissions and limitations under the License.
 
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./App";
-import * as serviceWorker from "./serviceWorker";
+import App from "./App.jsx";
+import * as serviceWorker from "./serviceWorker.js";
 
 ReactDOM.render(<App />, document.getElementById("root"));
