+++ sandbox/src/personalizationAnalyticsClientSideHelper.js
@@ -3,8 +3,8 @@ import {
   concatenateAnalyticsPayloads,
   getDisplayAnalyticsToken,
   getECID,
-} from "./analyticsTokenHandler";
-import { sendAnalyticsPayload } from "./DataInsertionAPI";
+} from "./analyticsTokenHandler.js";
+import { sendAnalyticsPayload } from "./dataInsertionApi.js";
 
 // this org is configured to have Analytics disabled thus it is A4T Client Side Logging
 const instanceName = "organizationTwo";
